[+] Externalization of the script: Institute
